create geometry store, start moving functions over
