+++ src/simularium/VisGeometry/types.ts
@@ -1,0 +1,8 @@
+import { Object3D } from "three";
+import { InstancedMesh } from "../rendering/InstancedMesh";
+
+export interface MeshLoadRequest {
+    mesh: Object3D;
+    cancelled: boolean;
+    instances: InstancedMesh;
+}
